Fixed a simulator crash while trying to play a song. Fixed crossfade when selecting a new track to play. Most likely fixed a bug which caused playback to stop on track change.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@7094 a1c6a512-1295-4272-9138-f99709370657

+++ apps/codecs.h
@@ -331,7 +331,7 @@ struct codec_api {
 
 /* defined by the codec loader (codec.c) */
 #if CONFIG_HWCODEC == MASNONE
-int codec_load_ram(char* codecptr, size_t size, void* ptr2, size_t bufwrap);
+int codec_load_ram(char* codecptr, int size, void* ptr2, int bufwrap);
 int codec_load_file(const char* codec);
 #endif
